Hadoop: Move exploit_host into Hadoop
