+++ monkey/tests/unit_tests/agent_plugins/exploiters/hadoop/test_hadoop.py
@@ -12,8 +12,10 @@ from agent_plugins.exploiters.hadoop.src.hadoop import Hadoop, HadoopOptions
 from common import OperatingSystem
 from common.agent_events import ExploitationEvent, PropagationEvent
 from common.event_queue import IAgentEventPublisher
+from infection_monkey.exploit import IAgentBinaryRepository
 from infection_monkey.model import TargetHost
 
+SERVERS = ["1.1.1.2, 1.1.1.3"]
 FAKE_AGENT_ID = UUID("9614480d-471b-4568-86b5-cb922a34ed8a")
 FAKE_APPLICATION_ID = "application_1404198295326_0003"
 FAKE_NEW_APPLICATION_RESPONSE_DICT = {
@@ -49,12 +51,26 @@ def interrupt():
 
 
 @pytest.fixture
+def mock_agent_binary_repository() -> IAgentBinaryRepository:
+    return MagicMock(spec=IAgentBinaryRepository)
+
+
+@pytest.fixture
 def mock_hadoop_exploiter(
-    mock_agent_event_publisher: IAgentEventPublisher, interrupt: Event
+    mock_agent_event_publisher: IAgentEventPublisher,
+    mock_agent_binary_repository: IAgentBinaryRepository,
+    interrupt: Event,
 ) -> Hadoop:
     target_host = TargetHost(ip=IPv4Address("1.1.1.1"), operating_system=OperatingSystem.WINDOWS)
     return Hadoop(
-        target_host, HadoopOptions(), interrupt, FAKE_AGENT_ID, mock_agent_event_publisher
+        target_host,
+        HadoopOptions(),
+        interrupt,
+        FAKE_AGENT_ID,
+        mock_agent_event_publisher,
+        mock_agent_binary_repository,
+        SERVERS,
+        current_depth=1,
     )
